requirements.txt
numpy
paho-mqtt
PyQt6
pyqtgraph
